+++ src/lens.js
@@ -1,5 +1,6 @@
 import _ from 'lodash/fp'
 import { setOn } from './conversion'
+import { toggleElementBy } from './array'
 
 // Stubs
 export let functionLens = val => (...x) => {
@@ -44,6 +45,12 @@ export let lensOf = object =>
     _.keys(object)
   )
 
+export let includeLens = (value, ...lens) => ({
+  get: () => _.includes(value, view(...lens)),
+  // Uniq is to ensure multiple calls to set(true) don't push multiple times since this is about membership of a set
+  set: x => set(_.uniq(toggleElementBy(!x, value, view(...lens))), ...lens),
+})
+
 // Lens Manipulation
 let construct = (...lens) => (lens[1] ? lensProp(...lens) : lens[0])
 let read = lens => (lens.get ? lens.get() : lens())
@@ -52,6 +59,35 @@ export let views = (...lens) => () => view(...lens)
 let write = (val, lens) => (lens.set ? lens.set(val) : lens(val))
 export let set = _.curryN(2, (val, ...lens) => write(val, construct(...lens)))
 export let sets = _.curryN(2, (val, ...lens) => () => set(val, ...lens))
+export let setsWith = _.curry((f, ...lens) => x =>
+  set(_.iteratee(f)(x), ...lens)
+)
 export let flip = (...lens) => () => set(!view(...lens), ...lens)
 export let on = sets(true)
 export let off = sets(false)
+
+// Lens Consumption
+// Map lens to dom event handlers
+let binding = (value, getEventValue) => (...lens) => ({
+  [value]: view(...lens),
+  onChange: setsWith(getEventValue, ...lens),
+})
+// Dom events have relevent fields on the `target` property of event objects
+let targetBinding = field => binding(field, `target.${field}`)
+export let domLens = {
+  value: targetBinding('value'),
+  checkboxValues: _.flow(
+    includeLens,
+    targetBinding('checked')
+  ),
+  hover: (...lens) => ({
+    onMouseOver: on(...lens),
+    onMouseOut: off(...lens),
+  }),
+  focus: (...lens) => ({
+    onFocus: on(...lens),
+    onBlur: off(...lens),
+  }),
+  targetBinding,
+  binding,
+}